Tarea #1455 - Nuevos campo segun posicion en cabecera de los documentos
